Move listeners from `App` to `ProxyLog`; Update `logMachine`
* Update `Chain`: add **TODO** (src/core/Chain/Chain.ts);
* Update `Logs`: fix **clear** method (src/core/Database/sublevels/Logs.ts);
* Update `App`: remove **updateLogs** & **createLogs** listeners (src/renderer/App.tsx);
* Update `ProxyLog`: add **updateLogs** & **createLogs** listeners; update **logGetAll** handler (src/renderer/components/templates/ProxyLog/ProxyLog.tsx);
* Update `logMachine`: add **add** event (src/xstate/logMachine.ts).
