+++ src/renderer/App.tsx
@@ -2,7 +2,6 @@ import { useEffect } from "react";
 import { uiActor, proxiesActor, logsActor } from "@/xstate";
 import Layout from "@/renderer/components/layout";
 import debug from "debug";
-import { ILogsRecord } from "@/types";
 debug.enable("*");
 
 const logger = debug("renderer:App");
@@ -20,42 +19,6 @@ const App = () => {
         logger(proxy);
       });
     });
-  }, []); // eslint-disable-line
-
-  useEffect(() => {
-    const removeEventListener = window.electronAPI.updateLogs(
-      (value: ILogsRecord) => {
-        logger(value);
-
-        logsActor.getSnapshot().context.logs[value.proxyId].send({
-          // TODO: remove `getSnapshot`
-          type: "update",
-          updatedLog: value,
-        });
-      }
-    );
-
-    return () => {
-      removeEventListener();
-    };
-  }, []); // eslint-disable-line
-
-  useEffect(() => {
-    const removeEventListener = window.electronAPI.createLogs(
-      (value: ILogsRecord) => {
-        logger(value);
-
-        logsActor.getSnapshot().context.logs[value.proxyId].send({
-          // TODO: remove `getSnapshot`
-          type: "create",
-          newLog: value,
-        });
-      }
-    );
-
-    return () => {
-      removeEventListener();
-    };
   }, []); // eslint-disable-line
 
   useEffect(() => {
