+++ src/renderer/components/templates/ProxyLog/ProxyLog.tsx
@@ -1,10 +1,11 @@
-import { Button, ScrollArea } from "@/renderer/components/ui";
+import { Button, ScrollArea, Typography } from "@/renderer/components/ui";
 import { uiActor, logsActor } from "@/xstate";
 import { XCircle, Trash2 } from "lucide-react";
 import { useSelector } from "@xstate/react";
 import debug from "debug";
 import { LogsRecord } from "@/renderer/components/templates";
 import { useEffect } from "react";
+import { ILogsRecord } from "@/types";
 const logger = debug("Renderer:ProxyLog");
 
 const ProxyLog = () => {
@@ -27,19 +28,56 @@ const ProxyLog = () => {
       for (const [key, value] of data) {
         logger("key:", key);
         logger("value:", value);
+
+        const { url, created, updated, stats } = value;
         logs.send({
-          // TODO: remove `getSnapshot`
-          type: "create",
-          newLog: {
+          type: "add",
+          log: {
             proxyId: logId as string,
             connectionId: Number(key),
-            url: value.url,
-            created: value.created,
-            updated: value.updated,
+            url,
+            created,
+            updated,
+            stats,
           },
         });
       }
     });
+  }, []); // eslint-disable-line
+
+  useEffect(() => {
+    const removeEventListener = window.electronAPI.updateLogs(
+      (value: ILogsRecord) => {
+        logger(value);
+
+        logs.send({
+          type: "update",
+          updatedLog: value,
+        });
+      }
+    );
+
+    return () => {
+      removeEventListener();
+    };
+  }, []); // eslint-disable-line
+
+  useEffect(() => {
+    const removeEventListener = window.electronAPI.createLogs(
+      (value: ILogsRecord) => {
+        logger(value);
+
+        logs.send({
+          // TODO: remove `getSnapshot`
+          type: "create",
+          newLog: value,
+        });
+      }
+    );
+
+    return () => {
+      removeEventListener();
+    };
   }, []); // eslint-disable-line
 
   return (
@@ -70,7 +108,7 @@ const ProxyLog = () => {
         </Button>
       )}
 
-      {logId}
+      <Typography variant="small">{logId}</Typography>
       {logId && (
         <ScrollArea className="h-full w-full rounded-md border p-4">
           <div className="flex flex-col gap-2">
